+++ agentrun-20250910/ChangeLog.md
@@ -1,3 +1,16 @@
+2025-10-21 Version: 2.0.0
+- Update API CreateMemory: add request parameters body.permanent.
+- Update API CreateMemory: add request parameters body.strategy.
+- Update API CreateMemory: add response parameters Body.data.
+- Update API GetMemory: add response parameters Body.data.cmsWorkspaceName.
+- Update API GetMemory: add response parameters Body.data.permanent.
+- Update API GetMemory: add response parameters Body.data.strategy.
+- Update API UpdateMemory: add request parameters body.
+- Update API UpdateMemory: add response parameters Body.data.
+- Update API UpdateMemory: delete request parameters longTtl.
+- Update API UpdateMemory: delete request parameters shortTtl.
+
+
 2025-10-16 Version: 1.1.0
 - Support API CreateMemory.
 - Support API CreateMemoryEvent.